chore: move inspector-panel folder into case-tile-common
