+++ v3/src/components/case-table/case-table-inspector.tsx
@@ -5,11 +5,11 @@ import InformationIcon from "../../assets/icons/icon-info.svg"
 import ScaleDataIcon from "../../assets/icons/icon-scaleData.svg"
 import TrashIcon from "../../assets/icons/icon-trash.svg"
 import ValuesIcon from "../../assets/icons/icon-values.svg"
-import { DatasetInfoModal } from "./inspector-panel/dataset-info-modal"
-import { TrashMenuList } from "./inspector-panel/trash-menu-list"
+import { DatasetInfoModal } from "../case-tile-common/inspector-panel/dataset-info-modal"
+import { TrashMenuList } from "../case-tile-common/inspector-panel/trash-menu-list"
 import { HideShowMenuList } from "../case-tile-common/hide-show-menu-list"
 import { t } from "../../utilities/translation/translate"
-import { RulerMenuList } from "./inspector-panel/ruler-menu-list"
+import { RulerMenuList } from "../case-tile-common/inspector-panel/ruler-menu-list"
 import { ITileInspectorPanelProps } from "../tiles/tile-base-props"
 import { ICaseTableModel, isCaseTableModel } from "./case-table-model"
 import { useDataSet } from "../../hooks/use-data-set"
