+++ v3/src/components/case-tile-common/inspector-panel/dataset-info-modal.scss
