+++ v3/src/components/case-tile-common/inspector-panel/trash-menu-list.tsx
